requirements.txt
Django==3.2.15
wheel==0.37.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
django-markdownx==3.0.1
Pygments==2.13.0
